# backend/oas3/backend-api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/users/login":{"post":{"tags":["user-controller"],"operationId":"login","requestBody":{"description":"the credentials of the user trying to log in.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCredentialsDTO"}}},"required":true},"responses":{"404":{"description":"No such user","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"200":{"description":"Success","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"/api/v1/users/create":{"post":{"tags":["user-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDTO"}}},"required":true},"responses":{"200":{"description":"Success","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"/api/v1/notifications/{userId}/remove":{"post":{"tags":["notification-controller"],"operationId":"removeNotificationsFor","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"ids","in":"query","required":true,"schema":{"type":"array","items":{"type":"integer","format":"int32"}}}],"responses":{"404":{"description":"No such user","content":{"*/*":{"schema":{"type":"boolean"}}}},"400":{"description":"Missing user id or notification ids","content":{"*/*":{"schema":{"type":"boolean"}}}},"200":{"description":"Success","content":{"*/*":{"schema":{"type":"boolean"}}}}}}},"/api/v1/notifications/push-to/{userId}":{"post":{"tags":["notification-controller"],"operationId":"pushToUser","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotificationDTO"}}},"required":true},"responses":{"400":{"description":"Missing user id","content":{"*/*":{"schema":{"type":"boolean"}}}},"404":{"description":"No such user","content":{"*/*":{"schema":{"type":"boolean"}}}},"200":{"description":"Success","content":{"*/*":{"schema":{"type":"boolean"}}}}}}},"/api/v1/notifications/push-to-group/{groupId}":{"post":{"tags":["notification-controller"],"operationId":"pushToGroup","parameters":[{"name":"groupId","in":"path","description":"id of group","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotificationDTO"}}},"required":true},"responses":{"404":{"description":"No such group","content":{"*/*":{"schema":{"type":"boolean"}}}},"400":{"description":"Missing group id","content":{"*/*":{"schema":{"type":"boolean"}}}},"200":{"description":"Success","content":{"*/*":{"schema":{"type":"boolean"}}}}}}},"/api/v1/groups/{groupId}/remove-user/{userInQuestion}":{"post":{"tags":["group-controller"],"operationId":"removeUser","parameters":[{"name":"userInQuestion","in":"path","description":"id of user to be remove","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"groupId","in":"path","description":"id of group","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"description":"Credentials of the user performing this action","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCredentialsDTO"}}},"required":true},"responses":{"400":{"description":"Invalid acting user","content":{"*/*":{"schema":{"type":"boolean"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"type":"boolean"}}}},"200":{"description":"Success","content":{"*/*":{"schema":{"type":"boolean"}}}}}}},"/api/v1/groups/{groupId}/delete":{"post":{"tags":["group-controller"],"operationId":"deleteGroup","parameters":[{"name":"groupId","in":"path","description":"group to delete","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"description":"of the user trying to delete said group","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCredentialsDTO"}}},"required":true},"responses":{"400":{"description":"Invalid acting user","content":{"*/*":{"schema":{"type":"boolean"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"type":"boolean"}}}},"200":{"description":"Success","content":{"*/*":{"schema":{"type":"boolean"}}}}}}},"/api/v1/groups/{groupId}/add-user/{userId}":{"post":{"tags":["group-controller"],"operationId":"joinGroup","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"groupId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Adds a user to an existing group","content":{"*/*":{"schema":{"type":"boolean"}}}}}}},"/api/v1/groups/join":{"post":{"tags":["group-controller"],"operationId":"linkJoin","parameters":[{"name":"groupId","in":"query","description":"id of group","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"userId","in":"query","description":"id of user","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"404":{"description":"No such user or no such group","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GroupDTO"}}}},"400":{"description":"Missing groupId or userId","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GroupDTO"}}}},"200":{"description":"Success","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GroupDTO"}}}}}}},"/api/v1/groups/create":{"post":{"tags":["group-controller"],"operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGroupDTO"}}},"required":true},"responses":{"200":{"description":"Success","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GroupDTO"}}}}}}},"/api/v1/users":{"get":{"tags":["user-controller"],"summary":"Example: /api/v1/users?ids=1,7,32,45","description":"Example: /api/v1/users?ids=1,7,32,45","operationId":"getUsers","parameters":[{"name":"ids","in":"query","required":true,"schema":{"type":"array","items":{"type":"integer","format":"int32"}}}],"responses":{"400":{"description":"Invalid id list","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDTO"}}}}},"200":{"description":"Success","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDTO"}}}}}}}},"/api/v1/users/{userId}":{"get":{"tags":["user-controller"],"operationId":"getUser","parameters":[{"name":"userId","in":"path","description":"id of user to get information of.","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"404":{"description":"No such user","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"200":{"description":"Success","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"/api/v1/users/{userId}/groups":{"get":{"tags":["user-controller"],"operationId":"getUserGroups","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Success","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GroupDTO"}}}}}}}},"/api/v1/notifications/{userId}":{"get":{"tags":["notification-controller"],"operationId":"getAllNotifications","parameters":[{"name":"userId","in":"path","description":"id of user","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"404":{"description":"No such user","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NotificationDTO"}}}}},"200":{"description":"Success","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NotificationDTO"}}}}}}}},"/api/v1/notifications/{userId}/amount":{"get":{"tags":["notification-controller"],"operationId":"getAmountFor","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"400":{"description":"Missing user id","content":{"*/*":{"schema":{"type":"integer","format":"int32"}}}},"404":{"description":"No such user","content":{"*/*":{"schema":{"type":"integer","format":"int32"}}}},"200":{"description":"Success","content":{"*/*":{"schema":{"type":"integer","format":"int32"}}}}}}},"/api/v1/groups/{groupId}":{"get":{"tags":["group-controller"],"operationId":"getGroup","parameters":[{"name":"groupId","in":"path","description":"id of group","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"404":{"description":"No such group","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GroupDTO"}}}},"400":{"description":"Missing group id","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GroupDTO"}}}},"200":{"description":"Success","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GroupDTO"}}}}}}}},"components":{"schemas":{"UserCredentialsDTO":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"description":""},"UserDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"username":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"phoneNumber":{"type":"string"}},"description":""},"CreateUserDTO":{"type":"object","properties":{"username":{"type":"string"},"passwordHash":{"type":"string"},"name":{"type":"string"},"phoneNumber":{"type":"string"},"email":{"type":"string"}},"description":""},"NotificationDTO":{"type":"object","properties":{"title":{"type":"string"},"message":{"type":"string"}},"description":""},"GroupDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"adminId":{"type":"integer","format":"int32"},"name":{"type":"string"},"descriptions":{"type":"string"},"users":{"type":"array","items":{"type":"integer","format":"int32"}}},"description":""},"CreateGroupDTO":{"type":"object","properties":{"idOfAdmin":{"type":"integer","format":"int32"},"name":{"type":"string"},"desc":{"type":"string"}},"description":""}}}}